% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_NN.R
\name{add_NN}
\alias{add_NN}
\title{add nearest neighbours}
\usage{
add_NN(se, res, colname = "nearest_reference")
}
\arguments{
\item{se}{SummarisedExperiment object}

\item{res}{the output of \code{\link[=ncla]{ncla()}}, the samples in the query column should
have identical names as in the \code{se} object}

\item{colname}{character (default \code{nearest_reference}), new colname to add to
the \code{colData()} slot}
}
\value{
a summarised experiment object with the nearest reference sample for
each query added to the colData slot
}
\description{
Add the nearest reference sample to the query data in a SummarizedExperiment
object.
}
