% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{batchCorrect}
\alias{batchCorrect}
\title{Batch correct}
\usage{
batchCorrect(x, y, method = "limma", prop.k = 0.05)
}
\arguments{
\item{x}{datasset one: sample x gene matrix}

\item{y}{datasset one: sample x gene matrix}

\item{method}{oneof c("MNN", "limma")}

\item{prop.k}{see batchelor::fastMNN}
}
\value{
legnth two list of batch corrected datasets
}
\description{
Wrapper for batch correction of two datasets
}
