% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd.R
\name{calcRCD}
\alias{calcRCD}
\title{Calculate RCD}
\usage{
calcRCD(x, g = NULL)
}
\arguments{
\item{x}{row matrix}

\item{g}{a character vector of feature names (colnames(x)) to use as atlas
genes for TSD}
}
\description{
Given a matrix x with samples in rows and genes in columns calculate the
Ranking Correlation Distance defined my Manatakis et al. 2021
}
\details{
Optionally give a subset of genes to use as the "atlas genes" as vector g
}
