% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd.R
\name{calcSRJSD}
\alias{calcSRJSD}
\title{Calculate srJSD}
\usage{
calcSRJSD(x, g = NULL)
}
\arguments{
\item{g}{}
}
\value{

}
\description{
Calculate square-root of the Jensen-Shannon divergence
for rows of matrix x
}
\details{
Optionally give a subset of genes to use as the "atlas genes" as vector g
}
