% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSD.R
\name{calcTSD}
\alias{calcTSD}
\title{Transcriptome Signature distance}
\usage{
calcTSD(x, g = NULL)
}
\arguments{
\item{g}{}
}
\value{

}
\description{
Given a gene expression matrix, x, with samples in rows calculate the
pairwise Transcriptome Signature Distances as per Manatakis et al. 2021
}
\details{
Optionally give a subset of genes to use as the "atlas genes" as vector g
}
