% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellLineUtils.R
\name{cleanCellID}
\alias{cleanCellID}
\title{clean cell IDs}
\usage{
cleanCellID(x, dataset)
}
\arguments{
\item{x}{character vector of cell IDs}

\item{dataset}{character denoting which cell line the IDs \code are from
One of
  - nci60Full
  - klijn2015Full
  - ccleFull}
}
\value{

}
\description{
given a character vector of cell-line IDs, return a standardised version.
}
\details{
all in UPPER case, no punctuation, and some other idiosyncrasies removed
}
