% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{dissim}
\alias{dissim}
\title{dissim}
\usage{
dissim(x, f, g = NULL)
}
\arguments{
\item{x}{row matrix}

\item{f}{dissimilarity measure to use; one of
  - "L1"
  - "L2"
  - "cosine"
  - "pearson"
  - "spearman"
  - "JSD"
  - "bhattacharyya"
  - "TSD"

  note: jensen-shannon and bhattacharyya distance use relative
  frequencies to calculate probability vectors and will
  fail if data contains negative values}
}
\value{
n x n dissimilarity matrix of x using function type f
}
\description{
dissim
}
