% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{.getBatchLimma}
\alias{.getBatchLimma}
\title{get batch}
\usage{
.getBatchLimma(df, batch)
}
\arguments{
\item{batch}{}
}
\value{

}
\description{
internal function to extract individual batch from limma corrected df
}
