% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{.getBatchMNN}
\alias{.getBatchMNN}
\title{get batch}
\usage{
.getBatchMNN(sce, batch)
}
\arguments{
\item{batch}{}
}
\value{

}
\description{
internal function to extract individual batch from MNN corrected df
}
