% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{.xdissim}
\alias{.xdissim}
\title{(internal) cross distance}
\usage{
.xdissim(x, y, f, g = NULL)
}
\arguments{
\item{x}{length p vector}

\item{y}{matrix with p cols}

\item{g}{}
}
\value{

}
\description{
(internal) cross distance
}
