% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{kns}
\alias{kns}
\title{k nearest samples}
\usage{
kns(xd, k = 1)
}
\arguments{
\item{k}{}
}
\value{
a `nrow(xd)` length list with k named + sorted elements per list
}
\description{
Given a matrix of cross distances, return the k nearest neighbours
for each of the queries (rows)
}
