% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_samples.R
\name{map_samples}
\alias{map_samples}
\title{map_samples}
\usage{
map_samples(query, reference, f)
}
\arguments{
\item{query}{row matrix with samples to query}

\item{reference}{row matrix with samples to use as the reference}

\item{f}{The dissimilarity function to use, see \code{\link[=dissim]{dissim()}}}
}
\value{
a data.frame
}
\description{
For a given set of \code{query} samples, perform dissimilarity mapping against the
supplied \code{reference} using the dissimilarity measure specified by \code{f}. This
function enforces that the features are the same, e.g. \code{query = x[, genes]}
and \code{reference = y[, genes]}.
}
\details{
Internally, this function uses foreach with the
ability to run in parallel, see \code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}} for how to
register a parallel backend.
}
