% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_samples.R
\name{map_samples_se}
\alias{map_samples_se}
\title{map_samples_se}
\usage{
map_samples_se(query, reference, f, i_q, i_r)
}
\arguments{
\item{query}{row matrix with samples to query}

\item{reference}{row matrix with samples to use as the reference}

\item{f}{The dissimilarity function to use, see \code{\link[=dissim]{dissim()}}}

\item{i_q}{The assay to use from the query SE}

\item{q_r}{The assay to use from the reference SE}
}
\value{
a data.frame
}
\description{
A light wrapper around \code{\link[=map_samples]{map_samples()}} for \link{SummarizedExperiment} objects
}
\details{
For a given set of \code{query} samples, perform dissimilarity mapping against the
supplied \code{reference} using the dissimilarity measure specified by \code{f}. This
function enforces that the features are the same, e.g. \code{query = x[, genes]}
and \code{reference = y[, genes]}.
}
