% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncla-package.R
\docType{package}
\name{ncla-package}
\alias{ncla}
\alias{ncla-package}
\title{ncla: What the Package Does (Title Case)}
\description{
Code for doing science with. more to come.
}
\details{
This package is designed to facilitate benchamrking of dissimilarity
measures for RNA-seq datasets
}
\author{
\strong{Maintainer}: Samuel Lee \email{lee.sa@wehi.edu.au}

}
\keyword{internal}
