% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncla.R
\name{ncla}
\alias{ncla}
\title{nearest cell line analysis}
\usage{
ncla(query, reference, f)
}
\arguments{
\item{query}{row matrix with samples to query}

\item{reference}{row matrix with samples to use as the reference}

\item{f}{The dissimilarity function to use, see `dissim()`}
}
\value{
a data.frame
}
\description{
For a given set of `query` samples, perform dissimilarity mapping against the
supplied `reference` using the dissimilarity measure specified by `f`. This
function enforces that the features are the same, e.g. `query = x[, genes]`
and `reference = y[, genes]`.
}
\details{
Internally, this function uses foreach with the
ability to run in parallel, see `doParallel::registerDoParallel()` for how to
register a parallel backend.
}
