% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGG_MDS_eda}
\alias{plotGG_MDS_eda}
\title{plot an MDS plot (ggplot edition)}
\usage{
plotGG_MDS_eda(dat, group, pal = NULL, title = "")
}
\arguments{
\item{dat}{MDS coordinates, the output of cmdscale(d, k=2)}

\item{title}{}
}
\value{

}
\description{
plot an MDS plot (ggplot edition)
}
