% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGG_PCA_eda}
\alias{plotGG_PCA_eda}
\title{plot PCA (ggplot edition)}
\usage{
plotGG_PCA_eda(dat, group, PCs = c(1, 2), pal = NULL, title = "")
}
\arguments{
\item{dat}{the PCA object: output of `prcomp`}

\item{group}{vector assigning samples from `dat` to groups to colour by}

\item{PCs}{vector of length two, integers specifying what PCs to plot}

\item{pal}{(default = null), character vector of colours the length of
the set of `group`. If left default this will be generated by using Set1
which has a max n of 9}

\item{title}{a title for the plot}
}
\description{
Make a PCA plot with specific group colourings for two select PCs
}
