% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGG_PCAscree_eda}
\alias{plotGG_PCAscree_eda}
\title{Plot PCA scree plot (ggplot edition)}
\usage{
plotGG_PCAscree_eda(dat, title = "")
}
\arguments{
\item{dat}{the PCA object: output of `prcomp`}

\item{title}{a title for the plot}
}
\description{
Plot PCA scree plot (ggplot edition)
}
