% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trom.R
\name{trom2}
\alias{trom2}
\title{Transcriptome Overlap Measure}
\usage{
trom2(x, z = 1.5)
}
\arguments{
\item{x}{a n x m data frame, normalised gene expression with \emph{n} samples in rows}

\item{z}{z-score threshold for associated genes}
}
\value{
n x n matrix of TROM similarity scores
}
\description{
Calculate pairwise TROM scores for a set of samples
}
\references{
Li, W.V., Chen, Y. & Li, J.J. Stat Biosci 9, 105–136 (2017).
}
