% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trom.R
\name{trom_ag}
\alias{trom_ag}
\title{TROM from associated genes}
\usage{
trom_ag(ag1, ag2, total, n_test = 1)
}
\arguments{
\item{ag1}{first list of associated genes}

\item{ag2}{second list of associated genes}

\item{total}{total number of genes in the dataset}

\item{n_test}{total number of tests for multiple testing correction}
}
\description{
given two sets of associated genes, calculate the TROM score
}
