% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{xdissim}
\alias{xdissim}
\title{cross distance}
\usage{
xdissim(x, y, f, g = NULL)
}
\arguments{
\item{x}{a row matrix (or coercable dataframe)}

\item{y}{a row matrix (or coercable dataframe)}

\item{f}{function to use, see dissim}

\item{g}{gene space for tsd, see dissim}
}
\value{

}
\description{
Given two matrices, x and y, find the distance between each row of x to each row of y
}
