% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trom.R
\name{calcAssociatedGenes}
\alias{calcAssociatedGenes}
\title{calcAssociatedGenes}
\usage{
calcAssociatedGenes(x, z = 1.5)
}
\arguments{
\item{x}{a n x m data frame, normalised gene expression with \emph{n} samples in rows}

\item{z}{z-score threshold for associated genes}
}
\value{
length(n) list with associated genes for each sample
}
\description{
given a data-set, z-score transform and get associated genes as per TROM method
}
\references{
Li, W.V., Chen, Y. & Li, J.J. Stat Biosci 9, 105–136 (2017).
}
