% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGG_MDS_eda}
\alias{plotGG_MDS_eda}
\title{plot an MDS plot (ggplot edition)}
\usage{
plotGG_MDS_eda(dat, group, shape = NULL, pal = NULL, title = "")
}
\arguments{
\item{dat}{MDS coordinates, the output of cmdscale(d, k=2)}

\item{group}{vector assigning samples from \code{dat} to groups to colour by}

\item{shape}{vector assigning samples from \code{dat} for different shapes}

\item{pal}{(default = null), character vector of colours the length of
the set of \code{group}. If left default this will be generated by using Set1
which has a max n of 9}

\item{title}{title of the plot}
}
\description{
plot an MDS plot (ggplot edition)
}
