% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGG_UMAP_eda}
\alias{plotGG_UMAP_eda}
\title{Plot UMAP (ggplot edition)}
\usage{
plotGG_UMAP_eda(dat, group, pal = NULL, title = "")
}
\arguments{
\item{title}{}
}
\value{

}
\description{
Plot UMAP (ggplot edition)
}
