% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{subsetGenes}
\alias{subsetGenes}
\title{Subset genes}
\usage{
subsetGenes(x, f)
}
\arguments{
\item{x}{row matrix where columns are ENSG labelled gens}

\item{f}{character name of geneset to use
- Halmarks
- L1000
- Stable}
}
\value{

}
\description{
subset a gene-expression matrix with a defined geneset
}
